[TM-2855] map code improvements
