+++ src/components/elements/Map-mapbox/interactions/draw.ts
@@ -3,11 +3,8 @@ import bbox from "@turf/bbox";
 import mapboxgl from "mapbox-gl";
 
 import { loadPolygonGeoJson, loadProjectPolygonsGeoJson } from "@/connections/GeoJsonExport";
-import { createProjectPolygonWithReplace, updateProjectPolygonResource } from "@/connections/ProjectPolygons";
-import { createSitePolygonsResource } from "@/connections/SitePolygons";
-import { CreateSitePolygonAttributesDto } from "@/generated/v3/researchService/researchServiceSchemas";
+import { updateProjectPolygonResource } from "@/connections/ProjectPolygons";
 import { GeoJsonExportDto } from "@/generated/v3/researchService/researchServiceSchemas";
-import ApiSlice from "@/store/apiSlice";
 import Log from "@/utils/log";
 
 import { zoomToBbox } from "../adapters/camera";
@@ -24,9 +21,6 @@ const extractGeoJsonFromResponse = (
     return (response as any).data.attributes as unknown as GeoJSON.FeatureCollection;
   return undefined;
 };
-
-export { convertToAcceptedGEOJSON } from "../adapters/geojson";
-export { convertToGeoJSON } from "../adapters/geojson";
 
 export const startDrawing = (draw: MapboxDraw, map: mapboxgl.Map): void => {
   draw.changeMode("draw_polygon");
@@ -134,76 +128,6 @@ export async function fetchPolygonGeometry(
   } catch (error) {
     Log.error("Failed to fetch polygon geometry:", error);
     throw error;
-  }
-}
-
-export async function storePolygon(
-  geojson: any,
-  record: any,
-  setPolygonFromMap?: any,
-  refetchSitePolygons?: () => any
-): Promise<void> {
-  if (!geojson?.length) return;
-
-  const attributes: CreateSitePolygonAttributesDto = {
-    geometries: [
-      {
-        type: "FeatureCollection",
-        features: [
-          {
-            type: "Feature",
-            geometry: geojson[0].geometry,
-            properties: { site_id: record.uuid }
-          }
-        ] as any
-      }
-    ]
-  };
-
-  try {
-    const result = await createSitePolygonsResource(attributes);
-    ApiSlice.pruneCache("boundingBoxes");
-    ApiSlice.pruneIndex("boundingBoxes", "");
-    if (refetchSitePolygons) await refetchSitePolygons();
-    if (setPolygonFromMap) {
-      setPolygonFromMap({ uuid: result.polygonUuid, isOpen: true, primary_uuid: result.primaryUuid });
-    }
-  } catch (error) {
-    console.error("Failed to create site polygon:", error);
-    throw error;
-  }
-}
-
-export async function storePolygonProject(
-  geojson: any,
-  entityUuid: string,
-  entityType: string,
-  refetch: any,
-  setPolygonFromMap: any
-): Promise<void> {
-  if (!geojson?.length) return;
-
-  const geometries = [
-    {
-      type: "FeatureCollection",
-      features: geojson.map((feature: any) => ({
-        type: "Feature",
-        geometry: feature.geometry,
-        properties: { projectPitchUuid: entityUuid }
-      }))
-    }
-  ];
-
-  const response = await createProjectPolygonWithReplace({ geometries }, entityUuid);
-  const polygonUuid = response.polygonUuid;
-  if (polygonUuid) {
-    refetch?.();
-    setPolygonFromMap?.({
-      uuid: polygonUuid,
-      isOpen: true,
-      entityName: "project-pitches",
-      projectPitchUuid: entityUuid
-    });
   }
 }
